// src/views/navigation/types/MenuItemCreate.ts
export interface MenuItemCreate_menuItemCreate_errors {
  __typename: "MenuError";
  code: MenuErrorCode;
  field: string | null;
}
// ...
export interface MenuItemCreate_menuItemCreate_menuItem {
  __typename: "MenuItem";
  menu: MenuItemCreate_menuItemCreate_menuItem_menu;
}

export interface MenuItemCreate_menuItemCreate {
  __typename: "MenuItemCreate";
  errors: MenuItemCreate_menuItemCreate_errors[];
  menuItem: MenuItemCreate_menuItemCreate_menuItem | null;
}

export interface MenuItemCreate {
  menuItemCreate: MenuItemCreate_menuItemCreate | null;
}